Revamp Rename edit detail product. Reallocate product views to productViews.py, and implement. Implement item templates to product. Delete product templates. Implement changes in urls. Reallocate version template to product folder.
