--- templates/product/detailEditProduct.html
@@ -1,15 +1,0 @@
-{% extends 'product/baseFrame.html' %}
-
-{% block cardTitle %}
-Edit
-{% endblock %}
-
-{% block content %}
-<form method="post" novalidate>
-    {% csrf_token %}
-    {% include 'include/form.html' %}
-    <div style="float: left;">
-        <button type="submit" class="btn btn-success">Save</button>
-    </div>
-</form>
-{% endblock %}
