--- templates/product/detailProduct.html
@@ -1,35 +1,0 @@
-{% extends 'product/baseFrame.html' %}
-
-{% block cardTitle %}
-Detail
-{% endblock %}
-
-{% block bodyParam %}
-onload="onLoad()"
-{% endblock %}
-
-{% block content %}
-<form method="post" novalidate>
-    {% csrf_token %}
-    {% include 'include/form.html' %}
-</form>
-<div style="float: left;">
-    <button class="btn btn-success" onclick="reload()">Edit</button>
-</div>
-{% endblock %}
-
-{% block script %}
-<script type="text/javascript">
-    function reload() {
-        window.location.href = '/detailEditProduct/{{ item.id }}';
-    }
-    function onLoad() {
-        var url = new URL(window.location.href);
-        var c = url.searchParams.get("page");
-        console.log(c);
-        if (c == "reload") {
-            parent.location.reload();
-        }
-    }
-</script>
-{% endblock %}
