--- templates/product/newProduct.html
@@ -1,26 +1,0 @@
-{% extends 'product/baseFrame.html' %}
-
-{% block cardTitle %}
-New
-{% endblock %}
-
-{% block content %}
-<form method="post" novalidate>
-    {% csrf_token %}
-    {% include 'include/form.html' %}
-    <div style="float: left;">
-        <button type="submit" class="btn btn-success">Add</button>
-    </div>
-    <div style="float: right;">
-        <button type="submit" class="btn btn-danger" onclick="reload()">Cancel</button>
-    </div>
-</form>
-{% endblock %}
-
-{% block script %}
-<script type="text/javascript">
-    function reload() {
-        parent.location.reload();
-    }
-</script>
-{% endblock %}
