--- templates/product/newProductReload.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-</head>
-<body onload="reload()">
-
-</body>
-</html>
-<script type="text/javascript">
-    function reload() {
-        parent.location.reload();
-    }
-</script>
